requirements.txt
Bootstrap-Flask==2.0.0
Flask==2.2.0
Flask-Testing==0.8.1
Flask-WTF==1.0.0
redis==4.1.1
rq==1.10.1
Werkzeug==2.2.2
Authlib
requests
PyGithub
PyJWT
flask-login
flask-sqlalchemy==3.0.3
wtforms_components
sphinx
flasgger==0.9.7b2
openapi3-parser
python-dotenv